requirements/dev.txt
fastapi
uvicorn
pydantic
boto3
pytest
httpx
python-dotenv
alembic
sqlalchemy
passlib[bcrypt]
pymysql
pyjwt
cryptography
gql[all]
python-multipart
pillow
oss2
pybase64
twilio
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
vonage==4.4.0
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.6
vonage-numbers==1.0.4
vonage-sms==1.1.5
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.3.0
cachetools==5.5.2
pycryptodome==3.23.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
aliyun-python-sdk-sts==3.1.0
alibabacloud-dkms-gcs==1.0.2
apscheduler==3.11.0
tenacity==9.1.2
psutil>=5.9.0
phonenumbers==9.0.14
langdetect>=1.0.0
openai>=1.0.0
# spacy>=3.7
alibabacloud_green20220302==2.20.3
presidio-analyzer>=2.2.0
